testing/web-platform/tests/workers/Worker_cross_origin_security_err.htm

<!DOCTYPE html>
<html>
<head>
<title> Worker SECURITY_ERR exception </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var description = "Test Description: Throw a SECURITY_ERR exception when origin of the "
                    + "worker script URL and origin URL are not same.";
                  
    test(function()
    {
        assert_throws("SECURITY_ERR", function()
        {
            var worker = new Worker("data://./support/WorkerBasic.js");
        });
    }, description);
</script>
</body>
</html>
